dev/sg: refactor sg teammate to use ffcli subcommands
